Arduino GPS Distance Code
This commit adds in a standalone Arduino sketch to use a GPS module and
calculate the arc distance between 2 GPS coordinates.  A pushbutton is
attached such that a user can walk around with the electronics and set
the desired start/end GPS coordinates.

The Haversine distance function is implemented to take into account the
GPS coordinates assume the Earth is an ellipsoid (see
https://www.themathdoctors.org/distances-on-earth-2-the-haversine-formula/).

Signed-off-by: Bryant Pong <rcxking@gmail.com>