Arduino Node GPS Upgrades
This commit adds in a 150 ms delay at the start of the setup() function
to avoid duplicating Serial.print() statements.  This is because the
Arduino Mega calls setup() twice when a Serial port is opened up.

Additionally, the GPS code has been optimized to remove the dependency
on the MicroNMEA library.

Signed-off-by: Bryant Pong <rcxking@gmail.com>